+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add standalone chart view from dashboard
+
 ### Changed
+- BREAKING: `charts` property is now a dictionary
+- BREAKING: removed `alias` property for charts
+- BREAKING: dashboard layout is now using chart key identifiers in the `charts` property
 - Set mobile layout breakpoint to 800px for better readability
 
 ## [0.1.3] - 2021-05-01